tuesday

+++ dominantColors.py
@@ -38,7 +38,8 @@ class DominantColors:
 
 
         #returning after converting to integer from float
-        return self.COLORS.astype(int)
+        # return self.COLORS.astype(int)
+        return self.COLORS
 
 
     def plotHistogram(self):
